% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_version.R
\name{api_version}
\alias{api_version}
\title{Set the API version.}
\usage{
api_version()
}
\value{
character. A string that sets the version for the API call to be used in all other functions.
}
\description{
Set the API version.
}
\examples{
api_version()

}
