% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{coin_categories}
\alias{coin_categories}
\title{Coin categories}
\usage{
coin_categories()
}
\value{
list.
}
\description{
List of coin categories. The function is run with no arguments and retreives the list from the API
}
\examples{
coin_categories()
}
