% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{coin_categories_data}
\alias{coin_categories_data}
\title{Coin categories with market data}
\usage{
coin_categories_data(order = "market_cap_desc")
}
\arguments{
\item{order}{character. See https://www.coingecko.com/en/api#explore-api for valid values.}
}
\value{
list.
}
\description{
List all categories with market data
}
\examples{
coin_categories_data()
}
