% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins_chart}
\alias{coins_chart}
\title{Charts}
\usage{
coins_chart(id, currency, days = 30)
}
\arguments{
\item{id}{Character. The id of the coin. refers to /coins/list.}

\item{currency}{Character. The target currency of market data (usd, eur, jpy, etc.)}

\item{days}{Integer. How many days of data to retreive.}
}
\value{
list.
}
\description{
Get historical market data include price, market cap, and 24h volume (granularity auto).Minutely data will be used for duration within 1 day, Hourly data will be used for duration between 1 day and 90 days, Daily data will be used for duration above 90 days.
}
