% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins_data}
\alias{coins_data}
\title{Title}
\usage{
coins_data(
  id = "bitcoin",
  localization = FALSE,
  tickers = FALSE,
  market_data = TRUE,
  community_data = FALSE,
  developer_data = FALSE,
  sparkline = FALSE
)
}
\arguments{
\item{id}{Character. The id of the coin. refers to /coins/list.}

\item{localization}{logical.}

\item{tickers}{logical.}

\item{market_data}{logical.}

\item{community_data}{logical.}

\item{developer_data}{logical.}

\item{sparkline}{logical.}
}
\value{
list.
}
\description{
Title
}
\examples{
coins_data()

}
