% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins_list}
\alias{coins_list}
\title{Coins list}
\usage{
coins_list(include_platform = FALSE)
}
\arguments{
\item{include_platform}{logical.}
}
\value{
list.
}
\description{
List all supported coins id, name and symbol.
}
\examples{
coins_list()
}
