% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_version.R
\name{get_json}
\alias{get_json}
\title{Make the call to the CoinGecko API}
\usage{
get_json(api_call)
}
\arguments{
\item{api_call}{character}
}
\value{
list
}
\description{
Make the call to the CoinGecko API
}
\examples{
get_json("https://api.coingecko.com/api/v3/ping")

}
