% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{markets}
\alias{markets}
\title{Markets}
\usage{
markets(
  vs_currency,
  ids,
  category = NA,
  order = "market_cap_desc",
  per_page = 10,
  page = 1,
  sparkline = FALSE,
  price_change_percentage = "24h"
)
}
\arguments{
\item{vs_currency}{Character. The target currency of market data (usd, eur, jpy, etc).}

\item{ids}{Character. The ids of the coin, comma separated crytocurrency symbols (base). refers to /coins/list.}

\item{category}{Character. filter by coin category. Refer to /coin/categories/list. Default=NA.}

\item{order}{Character. valid values: market_cap_desc, gecko_desc, gecko_asc, market_cap_asc, market_cap_desc, volume_asc, volume_desc, id_asc, id_desc}

\item{per_page}{Integer. valid values: 1…250. Total results per page}

\item{page}{Integer. Page through results}

\item{sparkline}{Logical. Include sparkline 7 days data (eg. true, false)}

\item{price_change_percentage}{Include price change percentage in 1h, 24h, 7d, 14d, 30d, 200d, 1y (eg. '1h,24h,7d' comma-separated, invalid values will be discarded)}
}
\value{
list.
}
\description{
Use this to obtain all the coins market data (price, market cap, volume)
}
\examples{
markets(vs_currency="usd", ids="bitcoin")
}
