% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{ping_gecko}
\alias{ping_gecko}
\title{Ping CoinGecko API
This function is run with no arguments.}
\usage{
ping_gecko()
}
\value{
character.
}
\description{
Ping CoinGecko API
This function is run with no arguments.
}
\examples{
ping_gecko()

}
