% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{price}
\alias{price}
\title{Title}
\usage{
price(
  ids,
  vs_currencies,
  include_market_cap = FALSE,
  include_24hr_vol = FALSE,
  include_24hr_change = FALSE,
  include_last_updated_at = FALSE
)
}
\arguments{
\item{ids}{id of coins. refers to \code{coins/list}. character}

\item{vs_currencies}{vs_currency of coins. refers to simple/supported_vs_currencies. character}

\item{include_market_cap}{logical.}

\item{include_24hr_vol}{logical.}

\item{include_24hr_change}{logical.}

\item{include_last_updated_at}{logical.}
}
\value{
list.
}
\description{
Title
}
\examples{
price(ids=c("bitcoin","ethereum"), vs_currencies=c("usd","aud"))
}
