% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{supported_vs_currencies}
\alias{supported_vs_currencies}
\title{Get list of supported_vs_currencies.
This function is run with no arguments.}
\usage{
supported_vs_currencies()
}
\value{
character. List of currency symbols.
}
\description{
Get list of supported_vs_currencies.
This function is run with no arguments.
}
\examples{
supported_vs_currencies()
}
