% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coins_tickers}
\alias{coins_tickers}
\title{Coin tickers}
\usage{
coins_tickers(id)
}
\arguments{
\item{id}{Character. The id of the coin. refers to /coins/list.}
}
\value{
list.
}
\description{
Coin tickers
}
\examples{
coins_tickers(id="bitcoin")
}
