% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coin_history.R
\name{get_coin_history}
\alias{get_coin_history}
\title{Get price history}
\usage{
get_coin_history(coin, date_POSIXlt, simplify = TRUE, localization = FALSE)
}
\arguments{
\item{coin}{character}

\item{date_POSIXlt}{POSIXlt}

\item{simplify}{logical}

\item{localization}{logical}
}
\value{
list
}
\description{
function retreives price data for specified coin on specified date
}
\examples{
get_coin_history(coin = "bitcoin", date_POSIXlt = as.POSIXlt("2015-12-01"))

}
